+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.6] - 2019-03-19
+### Added
+- Form example comment
+
+### Changed
+- `Form#complete!` renamed to `#submit` (#complete! removed)
+- `Validatable#validate!` removed
+
 ## [0.5.5] - 2019-03-13
 ### Added
 - LunaPark::Mappers::Simple